Add employee_schedule_viewer.py
This is a mockup program that I designed. The purpose of this is to simulate what the final schedule will look like for the employer. This is 1 of 2 possible ideas for schedule layouts I have, the other can be discussed if we see fit. Within the code itself, the example_input section will eventually receive input from the genetic algorithm and build schedules.